setup: added default (empty) production config files to install set, updated upgrade publish profile to exclude the default production config files so they don't get overwritten by an upgrade.
Updated linux installer and windows documentation to set permissions on the production config files so that Nucleus has write permissions.  This is so that the setup wizard and log files settings page can write to the configuration files.

Also updated linux and Azure setup guides to reflect changes in the setup wizard (automatic setup of database settings and file system provider settings)