fix: added bug workaround logic for Linux.
Resource monitoring (IResourceMonitor) is used by the system information page, and is also used by OpenTelemetry, but does not work properly in Ubuntu 24.04.2 because /sys/fs/cgroup/user.slice/memory.current contains 0 shortly after a restart, and LinuxUtilizationParserCgroupV2 throws an exception if the value is 0. Resource monitoring is initialized automatically during dependency injection setup, so this exception is un-catchable and crashes Nucleus during startup.

Because of the bug, in Ubuntu 24.04.2 instrumentation should not be enabled

AddResourceMonitoring should not be called unless instrumentation is enabled, or we are running in Windows.

The system administration page was updated to make IResourceMonitor an optional dependency, and to suppress the membory and CPU display if it is not available.